refactor: migrate the Events story to TypeScript
This facilitates maintenance and enables errors to be detected earlier.
Also display a description at the top of the Graph to help understand what the story is doing.
